fixes to style, app boilerplate and selenium testing

+++ frappe/utils/sel.py
@@ -7,42 +7,52 @@ from selenium.webdriver.common.keys import Keys
 from selenium import webdriver
 from selenium.webdriver.common.by import By
 from selenium.webdriver.support.ui import WebDriverWait
+from selenium.webdriver.support.select import Select
 from selenium.webdriver.support import expected_conditions as EC
 from selenium.common.exceptions import TimeoutException
 from urllib import unquote
-import time, frappe, subprocess
+import time, subprocess
+import signal
+import sys
 
-host = "http://localhost:8888"
+host = "http://localhost"
 pipe = None
-driver = None
-verbose = None
-host = None
+port = "8888"
+_driver = None
+_verbose = None
 logged_in = False
 cur_route = False
+input_wait = 0
 
-def start(_verbose=None):
-	global driver, verbose
-	verbose = _verbose
+def start(verbose=None, driver="PhantomJS"):
+	global _driver, _verbose
+	_verbose = verbose
 
-	start_test_server(verbose)
-	driver = webdriver.PhantomJS()
+	_driver = getattr(webdriver, driver)()
+
+	signal.signal(signal.SIGINT, signal_handler)
+
+def signal_handler(signal, frame):
+	close()
+	sys.exit(0)
 
 def start_test_server(verbose):
 	global pipe
-	pipe = subprocess.Popen(["frappe", "--serve", "--port", "8888"], stdout=subprocess.PIPE, stderr=subprocess.PIPE)
+	pipe = subprocess.Popen(["frappe", "--serve", "--port", port], stdout=subprocess.PIPE, stderr=subprocess.PIPE)
+	#time.sleep(5)
 	while not pipe.stderr.readline():
 		time.sleep(0.5)
 	if verbose:
 		print "Test server started"
 
 def get(url):
-	driver.get(url)
+	_driver.get(url)
 
 def login(wait_for_id="#page-desktop"):
 	global logged_in
 	if logged_in:
 		return
-	get(host + "/login")
+	get(host + ":" + port + "/login")
 	wait("#login_email")
 	set_input("#login_email", "Administrator")
 	set_input("#login_password", "admin" + Keys.RETURN)
@@ -87,12 +97,16 @@ def add_child(fieldname):
 def find(selector, everywhere=False):
 	if cur_route and not everywhere:
 		selector = cur_route + " " + selector
-	return driver.find_elements_by_css_selector(selector)
+	return _driver.find_elements_by_css_selector(selector)
 
 def set_field(fieldname, value):
 	set_input('input[data-fieldname="{0}"]'.format(fieldname), value + Keys.TAB)
 	wait_for_ajax()
-	time.sleep(0.5)
+
+def set_select(fieldname, value):
+	select = Select(find('select[data-fieldname="{0}"]'.format(fieldname))[0])
+	select.select_by_value(value)
+	wait_for_ajax()
 
 def primary_action():
 	find(".appframe-titlebar .btn-primary")[0].click()
@@ -117,9 +131,9 @@ def wait(selector, everywhere=False):
 	if cur_route and not everywhere:
 		selector = cur_route + " " + selector
 	try:
-		elem = WebDriverWait(driver, 10).until(
+		elem = WebDriverWait(_driver, 20).until(
 			EC.presence_of_element_located((By.CSS_SELECTOR, selector)))
-		if verbose:
+		if _verbose:
 			print "found " + selector
 	except TimeoutException:
 		print "not found " + selector
@@ -130,9 +144,14 @@ def set_input(selector, text):
 	elem = find(selector)[0]
 	elem.clear()
 	elem.send_keys(text)
+	if input_wait:
+		time.sleep(input_wait)
 
 def close():
-	global driver, pipe
-	driver.quit()
-	pipe.kill()
-	driver = pipe = None
+	global _driver, pipe
+	print "closing selenium"
+	if _driver:
+		_driver.quit()
+	if pipe:
+		pipe.kill()
+	_driver = pipe = None
